module template moved to src
