+++ src/modules/module-template/components/ExtensionComponent.ts
