+++ src/modules/module-template/hooks/afterRegistration.ts
