+++ src/modules/module-template/hooks/beforeRegistration.ts
