+++ src/modules/module-template/pages/ExtensionPage.vue
