+++ src/modules/module-template/queries/exampleQuery.ts
