+++ src/modules/module-template/router/afterEach.ts
