+++ src/modules/module-template/router/beforeEach.ts
