+++ src/modules/module-template/store/plugin.ts
