+++ src/modules/module-template/types/ExampleState.ts
